requirements.txt
clickhouse-driver==0.2.6
requests==2.30.0
pandas==2.0.1
scipy==1.10.1
scikit-learn==1.2.2

fastapi==0.95.1
numpy~=1.24.3
pymongo~=4.3.3
uvicorn==0.22.0
orjson==3.8.0
aiohttp==3.8.4
python-dotenv==1.0.0
asyncpg==0.27.0

pydantic~=1.10.7
aioredis==2.0.1
apscheduler==3.10.1
dropbox==11.36.0
gunicorn==20.1.0

